+++ src/components/cc-tcp-redirection-form/cc-tcp-redirection-form.js
@@ -7,14 +7,13 @@ import { i18n } from '../../lib/i18n.js';
 import { linkStyles } from '../../templates/cc-link/cc-link.js';
 
 const SKELETON_REDIRECTIONS = [
-  { namespace: 'default', sourcePort: 1234 },
-  { namespace: 'cleverapps' },
+  { state: 'loading' },
+  { state: 'loading' },
 ];
 
 /**
- * @typedef {import('./cc-tcp-redirection-form.types.js').ContextRedirectionType} ContextRedirectionType
- * @typedef {import('./cc-tcp-redirection-form.types.js').Redirection} Redirection
- * @typedef {import('./cc-tcp-redirection-form.types.js').RedirectionNamespace} RedirectionNamespace
+ * @typedef {import('./cc-tcp-redirection-form.types.js').TcpRedirectionFormContextType} TcpRedirectionFormContextType
+ * @typedef {import('./cc-tcp-redirection-form.types.js').TcpRedirectionFormState} TcpRedirectionFormState
  */
 
 /**
@@ -22,72 +21,78 @@ const SKELETON_REDIRECTIONS = [
  *
  * @cssdisplay block
  *
- * @event {CustomEvent<RedirectionNamespace>} cc-tcp-redirection:create - Fires a redirection namespace whenever the create button is clicked.
- * @event {CustomEvent<Redirection>} cc-tcp-redirection:delete - Fires a redirection whenever the delete button is clicked.
+ * @event {CustomEvent<CreateTcpRedirection>} cc-tcp-redirection:create - Fires a redirection namespace whenever the create button is clicked.
+ * @event {CustomEvent<DeleteTcpRedirection>} cc-tcp-redirection:delete - Fires a redirection whenever the delete button is clicked.
  */
 export class CcTcpRedirectionForm extends LitElement {
 
   static get properties () {
     return {
       context: { type: String },
-      error: { type: Boolean },
-      redirections: { type: Array },
+      redirections: { type: Object },
     };
   }
 
   constructor () {
     super();
 
-    /** @type {ContextRedirectionType} Defines in which context the form is used so it can show the appropriate description or lack thereof (defaults to user). */
+    /** @type {TcpRedirectionFormContextType} Defines in which context the form is used so it can show the appropriate description or lack thereof (defaults to user). */
     this.context = 'user';
 
-    /** @type {boolean} Sets a loading error state. */
-    this.error = false;
-
-    /** @type {Redirection[]|null} Sets the list of redirections. */
-    this.redirections = null;
-  }
-
-  _getRedirectionCount () {
-    if (this.context === 'admin' && this.redirections != null) {
-      const howManyRedirections = this.redirections.filter(({ sourcePort }) => sourcePort != null).length;
-      if (howManyRedirections >= 1) {
-        return html`<cc-badge circle weight="strong">${howManyRedirections}</cc-badge>`;
-      }
-    }
-    return '';
+    /** @type {TcpRedirectionFormState} Sets the list of redirections. */
+    this.redirections = { state: 'loading' };
   }
 
   render () {
-    const skeleton = (this.redirections == null);
-    const redirections = skeleton ? SKELETON_REDIRECTIONS : this.redirections;
+
+    const state = this.redirections.state;
     const blockState = (this.context === 'admin') ? 'close' : 'off';
 
     return html`
       <cc-block state="${blockState}">
-        <div slot="title">${i18n('cc-tcp-redirection-form.title')}${this._getRedirectionCount()}</div>
+
+        <div slot="title">
+          ${i18n('cc-tcp-redirection-form.title')}
+          ${this._renderRedirectionCountBadge()}
+        </div>
+
         ${this.context === 'user' ? html`
           <div class="description">${i18n('cc-tcp-redirection-form.description')}</div>
         ` : ''}
-        ${!this.error && redirections.length > 0 ? html`
-          ${redirections.map((redirection) => html`
-            <cc-tcp-redirection
-              namespace=${redirection.namespace}
-              .sourcePort=${redirection.sourcePort}
-              ?skeleton=${skeleton}
-              ?waiting=${redirection.waiting}
-              ?private=${redirection.private}
-            ></cc-tcp-redirection>
+
+        ${state === 'loading' ? html`
+          ${SKELETON_REDIRECTIONS.map((redirection) => html`
+            <cc-tcp-redirection .redirection=${redirection}></cc-tcp-redirection>
           `)}
         ` : ''}
-        ${!this.error && redirections.length === 0 ? html`
-          <div class="cc-block_empty-msg">${i18n('cc-tcp-redirection-form.empty')}</div>
+
+        ${state === 'loaded' ? html`
+          ${this.redirections.value.map((redirection) => html`
+            <cc-tcp-redirection .redirection=${redirection}></cc-tcp-redirection>
+          `)}
+          ${this.redirections.value.length === 0 ? html`
+            <div class="cc-block_empty-msg">${i18n('cc-tcp-redirection-form.empty')}</div>
+          ` : ''}
         ` : ''}
-        ${this.error ? html`
+
+        ${state === 'error' ? html`
           <cc-error>${i18n('cc-tcp-redirection-form.error')}</cc-error>
         ` : ''}
+
       </cc-block>
     `;
+  }
+
+  _renderRedirectionCountBadge () {
+    if (this.context === 'admin' && this.redirections.state === 'loaded') {
+      const redirectionCount = this.redirections.value.filter(({ sourcePort }) => sourcePort != null).length;
+      if (redirectionCount >= 1) {
+        return html`
+          <cc-badge circle weight="strong">${redirectionCount}</cc-badge>
+        `;
+      }
+    }
+    return '';
   }
 
   static get styles () {