+++ src/components/cc-tcp-redirection/cc-tcp-redirection.types.d.ts
@@ -1,10 +1,28 @@
-export interface RedirectionNamespace {
+interface TcpRedirection {
   namespace: string,
+  isPrivate: boolean,
+  sourcePort?: number,
 }
 
-export interface Redirection {
-  namespace: string,
-  sourcePort: number,
+export type TcpRedirectionState = TcpRedirectionStateLoading | TcpRedirectionStateLoaded | TcpRedirectionStateWaiting;
+
+interface TcpRedirectionStateLoading {
+  state: "loading";
 }
 
-export type ContextRedirectionType = "user" | "admin";
+interface TcpRedirectionStateLoaded extends TcpRedirection {
+  state: "loaded";
+}
+
+interface TcpRedirectionStateWaiting extends TcpRedirection {
+  state: "waiting";
+}
+
+interface CreateTcpRedirection {
+  namespace: string;
+}
+
+interface DeleteTcpRedirection {
+  namespace: string;
+  sourcePort: number;
+}