+++ CHANGELOG.md
@@ -79,6 +79,7 @@ title: Changelog
 * Add new dependency `@web/test-runner-commands` so that Web Test Runner can test both on desktop and mobile.
 * Add new helpers to extract stories and run accessibility test on them.
 * Add new test files in most of the component folders. These test files only contain accessibility tests for the moment.
+* `<cc-tcp-redirection-form>`: rework state, types and smart for TCP redirection components
 
 ## 9.0.0 (2022-07-19)